title screen to play screen (with quitting out of both cases
